enable choosing between aeon and pyeda for boolean function backend
- with boolfunclib argument
- with MPBN_BOOLFUNCLIB environment variable
